requirements.txt
Django==1.10.4
gunicorn==19.6.0
whitenoise==3.2
redis==2.10.5
python-twitter==3.2.1